rename reader to rss_reader
