+++ .travis.yml
@@ -4,9 +4,9 @@ python:
 install:
   - pip install -r requirements.txt
 script:
-  - cd reader
+  - cd rss_reader
   - pip install .
-  - nosetests --cover-branches --with-coverage --cover-min-percentage=50 --cover-package=reader --nologcapture
+  - nosetests --cover-branches --with-coverage --cover-min-percentage=50 --cover-package=rss_reader --nologcapture
   - pycodestyle --max-line-length=120 .
   - python ./../rss_reader_checker.py
   - cd -
