+++ rss_reader/rss_reader/cache_handlers.py
@@ -6,7 +6,7 @@ This module provides tools for working with cache
 import os
 import json
 import time
-from reader import str_funcs
+from rss_reader import str_funcs
 
 
 def get_feed_from_cache(date, limit):
