+++ rss_reader/tests/feed_test.py
@@ -5,7 +5,7 @@ import sys
 
 
 sys.path.append(os.path.dirname((os.path.dirname(__file__))))
-from reader.RssParser import RssParser
+from rss_reader.RssParser import RssParser
 
 
 class FeedTest(unittest.TestCase):
@@ -26,7 +26,7 @@ class FeedTest(unittest.TestCase):
     def test_empty_items(self):
         self.assertEqual(self.parser.items, [])
 
-    @patch('reader.RssParser.get_img')
+    @patch('rss_reader.RssParser.get_img')
     @patch('feedparser.parse')
     def test_feed_parser(self, parse_mock, get_img_mock):
         test_time = (2021, 5, 19, 21, 22, 56, 2, 139, 0)
